+++ docs/ChangeLog.md
@@ -290,6 +290,7 @@
 - main (`ble-update`): fix error message with system-wide installation of `ble.sh` (fixed by tars0x9752) 1d2a9c1 a450775
 - main. util: fix problems of readlink etc. found by test in macOS (reported by aiotter) `#D1849` fa955c1 `#D1855` a22e145
 - util (`ble/builtin/trap`): run EXIT trap in subshells `#D1862` XXXXXXX
+- util (`ble/builtin/trap`): fix the RETURN trap `#D1863` XXXXXXX
 
 ## Documentation